refactor(core): remove deprecated internal/core package
